+++ dist_node/parse/expression_parser.js
@@ -52,6 +52,7 @@ var __o = require("bennu")["parse"],
     punctuator = tokenParser["punctuator"],
     prefixedOp = tokenParser["prefixedOp"],
     identifier = value["identifier"],
+    identifierOrKeyword = value["identifierOrKeyword"],
     literal = value["literal"],
     operator = value["operator"],
     stringLiteral = value["stringLiteral"],
@@ -62,6 +63,9 @@ var __o = require("bennu")["parse"],
             functionExpression0 = __o5["functionExpression"];
         return functionExpression0;
     }));
+(expression = late((function() {
+    return expression;
+})));
 (expression = late((function() {
     return expression;
 })));
@@ -87,7 +91,7 @@ var args = label("Arguments", ((arg = expression), node(between(punctuator("("),
     "array element", arrayElement))), node(between(punctuator("["), punctuator("]"), eager(arrayElements)),
     ast_expression.ArrayExpression.create))));
 var propertyName, propertyInitializer, objectProperties;
-(objectLiteral = label("Object Literal", ((propertyName = choice(stringLiteral, numericLiteral, identifier)), (
+(objectLiteral = label("Object Literal", ((propertyName = choice(stringLiteral, numericLiteral, identifierOrKeyword)), (
     propertyInitializer = label("Property Initializer", nodea(enumeration(then(propertyName, punctuator(":")),
         expression), ast_value.ObjectValue.create))), (objectProperties = sepBy(punctuator(","), expected(
     "object property", propertyInitializer))), node(between(punctuator("{"), punctuator("}"), eager(
@@ -147,21 +151,22 @@ var argumentList = label("Argument List", either(attempt(node(operatorExpression
         expr) {
         return ast_expression.OperatorExpression.create(expr.loc, expr);
     })), between(punctuator("("), punctuator(")"), expected("expression", expression)))));
-var accessorTarget = either(identifier.map((function(x0) {
+var accessorTarget = either(identifierOrKeyword.map((function(x0) {
         return [x0, false];
     })), between(punctuator("("), punctuator(")"), expected("accessor expression", expression))
     .map((function(x0) {
         return [x0, true];
     })));
-(accessor = label("Accessor", node(next(punctuator("."), accessorTarget), (function(loc, __o5) {
-    var x0 = __o5[0],
-        computed = __o5[1];
-    return ({
-        "loc": loc,
-        "property": x0,
-        "computed": computed
-    });
-}))));
+(accessor = label("Accessor", next(punctuator("."), accessorTarget)
+    .map((function(__o5) {
+        var x0 = __o5[0],
+            computed = __o5[1];
+        return ({
+            "loc": x0.loc,
+            "property": x0,
+            "computed": computed
+        });
+    }))));
 (newExpression = label("New Expression", nodea(next(keyword("new"), enumeration(expected("new object", memberExpression),
         expected("argument list", either(args, eager(enumeration(curryExpression)))))), ast_expression.NewExpression
     .create)));
@@ -176,7 +181,7 @@ var accessorReducer = (function(p, c) {
 })))));
 var leftHandSideExpression = label("Call Expression", ((reducer0 = (function(p, c) {
     return (c.hasOwnProperty("property") ? ast_expression.MemberExpression.create(SourceLocation.merge(
-        p.loc, c.loc), p, c.property, c.computed, c.checked) : ast_expression.CallExpression.create(
+        p.loc, c.loc), p, c.property, c.computed) : ast_expression.CallExpression.create(
         SourceLocation.merge(p.loc, c.loc), p, c));
 })), binds(enumeration(memberExpression, many(either(argumentList, accessor))), ((x1 = foldl.bind(null,
     reducer0)), (function() {