+++ dist_node/parse/token_parser.js
@@ -4,22 +4,19 @@
 */
 "use strict";
 var __o = require("bennu")["parse"],
-    punctuator, anyPunctuator, operator, anyOperator, keyword, identifier, anyIdentifier, nullLiteral, booleanLiteral,
-        numericLiteral, stringLiteral, regularExpressionLiteral, prefixedOp, prefixOperator, infixOperator,
-        unaryOperator, binaryOperator, ExpectError = __o["ExpectError"],
+    punctuator, anyPunctuator, operator, anyOperator, keyword, anyKeyword, identifier, anyIdentifier, nullLiteral,
+        booleanLiteral, numericLiteral, stringLiteral, regularExpressionLiteral, prefixedOp, prefixOperator,
+        infixOperator, unaryOperator, binaryOperator, ExpectError = __o["ExpectError"],
     token = __o["token"],
     indexOf = Function.prototype.call.bind(Array.prototype.indexOf),
-    join = Function.prototype.call.bind(Array.prototype.join),
-    expectError = (function(msg) {
-        return (function(pos, tok) {
-            return new(ExpectError)(pos, msg, (tok ? tok.value : "end of input"));
-        });
-    });
+    join = Function.prototype.call.bind(Array.prototype.join);
 (punctuator = (function() {
-    var options = arguments;
+    var msg, options = arguments;
     return token((function(tok) {
         return ((tok.type === "Punctuator") && (indexOf(options, tok.value) >= 0));
-    }), expectError(join(options, ", ")));
+    }), ((msg = join(options, ", ")), (function(pos, tok) {
+        return new(ExpectError)(pos, msg, (tok ? tok.value : "end of input"));
+    })));
 }));
 (anyPunctuator = token((function(tok) {
     return (tok.type === "Punctuator");
@@ -27,10 +24,12 @@ var __o = require("bennu")["parse"],
     return new(ExpectError)(pos, "any punctuator", (tok ? tok.value : "end of input"));
 })));
 (operator = (function() {
-    var options = arguments;
+    var msg, options = arguments;
     return token((function(tok) {
         return ((tok.type === "Operator") && (indexOf(options, tok.value) >= 0));
-    }), expectError(join(options, ", ")));
+    }), ((msg = join(options, ", ")), (function(pos, tok) {
+        return new(ExpectError)(pos, msg, (tok ? tok.value : "end of input"));
+    })));
 }));
 (anyOperator = token((function(tok) {
     return (tok.type === "Operator");
@@ -38,22 +37,31 @@ var __o = require("bennu")["parse"],
     return new(ExpectError)(pos, "any operator", (tok ? tok.value : "end of input"));
 })));
 (keyword = (function() {
-    var options = arguments;
+    var msg, options = arguments;
     return token((function(tok) {
         return ((tok.type === "Keyword") && (indexOf(options, tok.value) >= 0));
-    }), expectError(join(options, ", ")));
+    }), ((msg = join(options, ", ")), (function(pos, tok) {
+        return new(ExpectError)(pos, msg, (tok ? tok.value : "end of input"));
+    })));
+}));
+(anyKeyword = token((function(tok) {
+    return (tok.type === "Keyword");
+}), (function(pos, tok) {
+    return new(ExpectError)(pos, "any keyword", (tok ? tok.value : "end of input"));
+})));
+(identifier = (function() {
+    var msg, options = arguments;
+    return token((function(tok) {
+        return ((tok.type === "Identifier") && (indexOf(options, tok.value) >= 0));
+    }), ((msg = join(options, ", ")), (function(pos, tok) {
+        return new(ExpectError)(pos, msg, (tok ? tok.value : "end of input"));
+    })));
 }));
 (anyIdentifier = token((function(tok) {
     return (tok.type === "Identifier");
 }), (function(pos, tok) {
     return new(ExpectError)(pos, "any identifier", (tok ? tok.value : "end of input"));
 })));
-(identifier = (function() {
-    var options = arguments;
-    return token((function(tok) {
-        return ((tok.type === "Identifier") && (indexOf(options, tok.value) >= 0));
-    }), expectError(join(options, ", ")));
-}));
 (nullLiteral = token((function(tok) {
     return (tok.type === "Null");
 }), (function(pos, tok) {
@@ -86,16 +94,20 @@ var __o = require("bennu")["parse"],
     }));
 }));
 (prefixOperator = (function() {
-    var options = arguments;
+    var msg, options = arguments;
     return token((function(tok) {
         return ((tok.type === "PrefixOperator") && (indexOf(options, tok.value) >= 0));
-    }), expectError(join(options, ", ")));
+    }), ((msg = join(options, ", ")), (function(pos, tok) {
+        return new(ExpectError)(pos, msg, (tok ? tok.value : "end of input"));
+    })));
 }));
 (infixOperator = (function() {
-    var options = arguments;
+    var msg, options = arguments;
     return token((function(tok) {
         return ((tok.type === "InfixOperator") && (indexOf(options, tok.value) >= 0));
-    }), expectError(join(options, ", ")));
+    }), ((msg = join(options, ", ")), (function(pos, tok) {
+        return new(ExpectError)(pos, msg, (tok ? tok.value : "end of input"));
+    })));
 }));
 (unaryOperator = token((function(tok) {
     return (tok.type === "PrefixOperator");
@@ -112,6 +124,7 @@ var __o = require("bennu")["parse"],
 (exports["operator"] = operator);
 (exports["anyOperator"] = anyOperator);
 (exports["keyword"] = keyword);
+(exports["anyKeyword"] = anyKeyword);
 (exports["identifier"] = identifier);
 (exports["anyIdentifier"] = anyIdentifier);
 (exports["nullLiteral"] = nullLiteral);