+++ tests/parse/function.js
@@ -105,3 +105,4 @@ exports.with_statement_body = function(test) {
 
     test.done();
 };
+